Redesigned clay.io loading mechanism due to socket.io problems.
clay.io loads socket.io which occasionally goes in conflict with
require.js and produces a 'mismatched anonymous define()' error. This
seems to happen because socket.io spits an anonymous define() call when
included. This most likely happens because socket.io is not loaded by
our require.js.
This fix actually waits for clay.io and all it's required libs to be
loaded and then injects the require.js <script> tag in the DOM which
triggers the game load mechanism. This has one nasty side effect that if
clay.io is not available the game would not load :( !